Cepstrum bugs

+++ LPC/Cepstrum.h
@@ -57,9 +57,9 @@ Thing_define (PowerCepstrum, Cepstrum) {
 	ny = 1
 */
 
-autoCepstrum Cepstrum_create (double qmax, long nq);
+autoCepstrum Cepstrum_create (double qmax, integer nq);
 
-autoPowerCepstrum PowerCepstrum_create (double qmax, long nq);
+autoPowerCepstrum PowerCepstrum_create (double qmax, integer nq);
 /* Preconditions:
 		nq >= 2;
 	Postconditions:
@@ -99,13 +99,13 @@ void PowerCepstrum_fitTiltLine (PowerCepstrum me, double qmin, double qmax, doub
 autoPowerCepstrum PowerCepstrum_subtractTilt (PowerCepstrum me, double qstartFit, double qendFit, int lineType, int fitMethod);
 void PowerCepstrum_subtractTilt_inplace (PowerCepstrum me, double qstartFit, double qendFit, int lineType, int fitMethod);
 
-void PowerCepstrum_smooth_inplace (PowerCepstrum me, double quefrencyAveragingWindow, long numberOfIterations);
-autoPowerCepstrum PowerCepstrum_smooth (PowerCepstrum me, double quefrencyAveragingWindow, long numberOfIterations);
+void PowerCepstrum_smooth_inplace (PowerCepstrum me, double quefrencyAveragingWindow, integer numberOfIterations);
+autoPowerCepstrum PowerCepstrum_smooth (PowerCepstrum me, double quefrencyAveragingWindow, integer numberOfIterations);
 
 autoMatrix PowerCepstrum_to_Matrix (PowerCepstrum me);
 autoPowerCepstrum Matrix_to_PowerCepstrum (Matrix me);
-autoPowerCepstrum Matrix_to_PowerCepstrum_row (Matrix me, long row);
-autoPowerCepstrum Matrix_to_PowerCepstrum_column (Matrix me, long col);
+autoPowerCepstrum Matrix_to_PowerCepstrum_row (Matrix me, integer row);
+autoPowerCepstrum Matrix_to_PowerCepstrum_column (Matrix me, integer col);
 autoPowerCepstrum Cepstrum_downto_PowerCepstrum (Cepstrum me);
 
 #endif /* _Cepstrum_h_ */
